[SCHEDULE] First working pass at a schedule not accounting for dates

+++ AmortisedInterestSchedule/Rate.fs
@@ -3,6 +3,6 @@ module AmortisedInterestSchedule.Rate
 type RateFrequency = Monthly | Yearly
 
 type Rate = {
-    Amount: float
+    Amount: double
     Frequency: RateFrequency
 }
